NSKeyedArchiver, NSKeyedUnarchiver now use Any instead of AnyObject.
This also fixes the NSKeyedArchiver tests for macOS, which were probably
failing because of assumptions made in the implementation about
AnyObject. To help reduce the amount of casting in the implementation,
this commit adds a private NS "toll free bridged" type for the private
CF type used as a keyed archiver UID in CFPropertyList.